+++ src/sys/signal.rs
@@ -612,6 +612,14 @@ impl From<Signal> for SigSet {
     }
 }
 
+impl std::ops::BitOr for SigSet {
+    type Output = Self;
+
+    fn bitor(self, rhs: Self) -> Self::Output {
+        self.iter().chain(rhs.iter()).collect()
+    }
+}
+
 impl AsRef<libc::sigset_t> for SigSet {
     fn as_ref(&self) -> &libc::sigset_t {
         &self.sigset